Remove "EventStreamRpc" prefix from class names.
